I worked on problem 2, tidying the data by using a pivot wider to convert all the individual minute time counts to instead consist of a time variable and a activity variable that corresponds to that specific time period.
I then created a table that consisted of the total activity variable for each day by selecting the time points that corresponded for the unique day and week and then added up all activity times for those unique day and weekend time points. I then expanded the data so that there was a variabel for week that consisted of the unique day and total activity so that it was easier to read by the reader.